refactor: consumer
